chore(tasks): move CommentsSetupProvider to core
