+++ packages/sanity/src/core/studio/commentsSetup/index.ts
@@ -1,3 +1,4 @@
 export * from './CommentsSetupContext'
 export * from './CommentsSetupProvider'
 export * from './types'
+export * from './useCommentsSetup'
