+++ packages/sanity/src/core/studio/commentsSetup/types.ts
@@ -1,0 +1,17 @@
+import {type SanityClient} from '@sanity/client'
+
+/**
+ * @beta
+ * @hidden
+ */
+export interface CommentsSetupContextValue {
+  /**
+   * Addon dataset client, currently called `comments` dataset.
+   */
+  client: SanityClient | null
+  isRunningSetup: boolean
+  /**
+   * Function to create the addon dataset if it does not exist.
+   */
+  runSetup: () => Promise<SanityClient | null>
+}
