+++ packages/sanity/src/structure/comments/src/context/index.ts
@@ -3,5 +3,4 @@ export * from './enabled'
 export * from './intent'
 export * from './onboarding'
 export * from './selected-path'
-export * from './setup'
 export * from './upsell'
