+++ packages/sanity/src/structure/comments/src/hooks/index.ts
@@ -1,9 +1,9 @@
+export * from '../../../../core/studio/commentsSetup/useCommentsSetup'
 export * from './use-comment-operations'
 export * from './useComments'
 export * from './useCommentsEnabled'
 export * from './useCommentsIntent'
 export * from './useCommentsOnboarding'
 export * from './useCommentsSelectedPath'
-export * from './useCommentsSetup'
 export * from './useCommentsUpsell'
 export * from './useResolveCommentsEnabled'
