+++ packages/sanity/src/structure/comments/src/hooks/use-comment-operations/createOperation.ts
@@ -25,7 +25,7 @@ interface CreateOperationProps {
   onCreate?: (comment: CommentPostPayload) => void
   onCreateError: (id: string, error: Error) => void
   projectId: string
-  runSetup: (comment: CommentPostPayload) => Promise<void>
+  runSetup: () => Promise<SanityClient | null>
   workspace: string
 }
 
@@ -118,7 +118,11 @@ export async function createOperation(props: CreateOperationProps): Promise<void
   // comment to create.
   if (!client) {
     try {
-      await runSetup(nextComment)
+      const newAddonClient = await runSetup()
+      if (!newAddonClient) {
+        throw new Error('Failed to create addon dataset client')
+      }
+      await newAddonClient.create(nextComment)
     } catch (err) {
       onCreateError?.(nextComment._id, err)
       throw err
